% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspec.R
\name{SSNoiseReduction}
\alias{SSNoiseReduction}
\title{SSNoiseReduction: Automatic Noise Reduction using Spectral Subtraction}
\usage{
SSNoiseReduction(x, windowWidth = 0.02, alpha = 1, beta = 0.001)
}
\arguments{
\item{x}{a Wave object}

\item{windowWidth}{width of the analysis window (in sec)}

\item{alpha}{alpha parameter of SS}

\item{beta}{beta parameter of SS}
}
\value{
a Wave object of noise-reduced signal
}
\description{
SSNoiseReduction: Automatic Noise Reduction using Spectral Subtraction
}
